+++ packages/jest-circus/runner.js
@@ -8,3 +8,8 @@
 // Allow people to use `jest-circus/runner` as a runner.
 const runner = require('./build/legacy-code-todo-rewrite/jestAdapter');
 module.exports = runner;
+
+exports = require('./build/state');
+
+const run = require('./build/run');
+exports.run = run;